#37 rename model files
