+++ packages/api/src/model/error.model.ts
