+++ packages/api/src/model/index.ts
@@ -1,6 +1,6 @@
-export * from './auth'
-export * from './context'
-export * from './entity'
-export * from './task'
-export * from './user'
-export * from './error'
+export * from './auth.model'
+export * from './context.model'
+export * from './entity.model'
+export * from './task.model'
+export * from './user.model'
+export * from './error.model'
